tag parser cleanup
