+++ src/extract.cc
@@ -4,13 +4,11 @@
 #include <windows.h>
 #endif
 
-#include <cctype>
-#include <cmath>
 #include <cstdint>
-#include <cstdlib>
 #include <algorithm>
 #include <iostream>
 #include <limits>
+#include <memory>
 #include <optional>
 #include <string>
 
@@ -40,6 +38,7 @@
 
 #include "osr/elevation_storage.h"
 #include "osr/extract/conditional_parser.h"
+#include "osr/extract/tag_parser.h"
 #include "osr/extract/tags.h"
 #include "osr/lookup.h"
 #include "osr/platforms.h"
@@ -76,20 +75,6 @@ bool is_number(std::string_view s) {
          utl::all_of(s, [](char const c) { return std::isdigit(c); });
 }
 
-bool is_space(char const c) {
-  return std::isspace(static_cast<unsigned char>(c)) != 0;
-}
-
-std::string_view trim(std::string_view value) {
-  while (!value.empty() && is_space(value.front())) {
-    value.remove_prefix(1U);
-  }
-  while (!value.empty() && is_space(value.back())) {
-    value.remove_suffix(1U);
-  }
-  return value;
-}
-
 std::optional<resolved_restriction::type> parse_turn_restriction_type(
     std::string_view const value) {
   if (value.starts_with("no"sv)) {
@@ -114,128 +99,6 @@ std::optional<conditional_turn_restriction> parse_conditional_turn_restriction(
     return std::nullopt;
   }
   return conditional_turn_restriction{*type, condition};
-}
-
-struct osm_measure {
-  double value_{};
-  std::string_view unit_;
-};
-
-std::optional<osm_measure> parse_measure(std::string_view value) {
-  value = trim(value);
-  if (value.empty()) {
-    return std::nullopt;
-  }
-
-  auto* end = static_cast<char*>(nullptr);
-  auto const parsed_value = std::strtod(value.data(), &end);
-  if (end == value.data() || !std::isfinite(parsed_value)) {
-    return std::nullopt;
-  }
-
-  auto const unit_pos = static_cast<std::size_t>(end - value.data());
-  return osm_measure{parsed_value, trim(value.substr(unit_pos))};
-}
-
-std::optional<double> parse_length_m(std::string_view value) {
-  auto const measure = parse_measure(value);
-  if (!measure.has_value()) {
-    return std::nullopt;
-  }
-
-  auto unit = trim(measure->unit_);
-  if (unit.starts_with('\'')) {
-    unit.remove_prefix(1U);
-    if (auto const inches = parse_measure(unit);
-        inches.has_value() && trim(inches->unit_) == "\""sv) {
-      return (measure->value_ * 12.0 + inches->value_) * 0.0254;
-    }
-    if (unit.empty()) {
-      return measure->value_ * 0.3048;
-    }
-  }
-
-  switch (cista::hash(unit)) {
-    case cista::hash(""):
-    case cista::hash("m"):
-    case cista::hash("metre"):
-    case cista::hash("metres"):
-    case cista::hash("meter"):
-    case cista::hash("meters"): return measure->value_;
-    case cista::hash("km"):
-    case cista::hash("kilometre"):
-    case cista::hash("kilometres"):
-    case cista::hash("kilometer"):
-    case cista::hash("kilometers"): return measure->value_ * 1000.0;
-    case cista::hash("mi"):
-    case cista::hash("mile"):
-    case cista::hash("miles"): return measure->value_ * 1609.344;
-    case cista::hash("nmi"): return measure->value_ * 1852.0;
-    case cista::hash("yd"):
-    case cista::hash("yds"): return measure->value_ * 0.9144;
-    case cista::hash("ft"): return measure->value_ * 0.3048;
-    case cista::hash("in"):
-    case cista::hash("\""): return measure->value_ * 0.0254;
-    default: return std::nullopt;
-  }
-}
-
-std::optional<double> parse_weight_t(std::string_view value) {
-  auto const measure = parse_measure(value);
-  if (!measure.has_value()) {
-    return std::nullopt;
-  }
-
-  auto const unit = trim(measure->unit_);
-  switch (cista::hash(unit)) {
-    case cista::hash(""):
-    case cista::hash("t"): return measure->value_;
-    case cista::hash("kg"): return measure->value_ / 1000.0;
-    case cista::hash("st"):
-    case cista::hash("ton"):
-    case cista::hash("tons"): return measure->value_ * 0.9071847;
-    case cista::hash("lt"): return measure->value_ * 1.016047;
-    case cista::hash("lb"):
-    case cista::hash("lbs"): return measure->value_ * 0.00045359237;
-    case cista::hash("cwt"): return measure->value_ * 0.0508;
-    default: return std::nullopt;
-  }
-}
-
-std::optional<double> parse_speed_km_h(std::string_view value) {
-  auto const measure = parse_measure(value);
-  if (!measure.has_value()) {
-    return std::nullopt;
-  }
-
-  switch (cista::hash(measure->unit_)) {
-    case cista::hash(""):
-    case cista::hash("km/h"):
-    case cista::hash("kph"):
-    case cista::hash("kmh"):
-    case cista::hash("kmph"): return measure->value_;
-    case cista::hash("mph"): return measure->value_ * 1.609344;
-    case cista::hash("knots"): return measure->value_ * 1.852;
-    default: return std::nullopt;
-  }
-}
-
-std::optional<double> parse_unitless(std::string_view value) {
-  auto const measure = parse_measure(value);
-  if (!measure.has_value() || !trim(measure->unit_).empty()) {
-    return std::nullopt;
-  }
-  return measure->value_;
-}
-
-template <typename T>
-std::optional<T> to_integer(std::optional<double> const value,
-                            double const factor = 1.0) {
-  if (!value.has_value() || *value < 0.0) {
-    return std::nullopt;
-  }
-  return static_cast<T>(std::min<double>(std::round(*value * factor),
-                                         std::numeric_limits<T>::max()));
 }
 
 template <typename T, typename Value>
