+++ test/conditional_parser_test.cc
@@ -374,6 +374,9 @@ TEST(conditional_parser, parses_supported_numeric_restrictions_and_units) {
            test_case{"maxlength:hgv:conditional"sv, "2 mi"sv,
                      osr::conditional_restriction_field::kMaxLength, 321869U,
                      osr::conditional_numeric_unit::kCentimeter},
+           test_case{"maxlength:hgv:conditional"sv, "2 nmi"sv,
+                     osr::conditional_restriction_field::kMaxLength, 370400U,
+                     osr::conditional_numeric_unit::kCentimeter},
            test_case{"maxheight:hgv:backward:conditional"sv, "12'6\""sv,
                      osr::conditional_restriction_field::kMaxHeight, 381U,
                      osr::conditional_numeric_unit::kCentimeter},
@@ -382,6 +385,9 @@ TEST(conditional_parser, parses_supported_numeric_restrictions_and_units) {
                      osr::conditional_numeric_unit::kCentimeter},
            test_case{"maxweight:hgv:conditional"sv, "10000 kg"sv,
                      osr::conditional_restriction_field::kMaxWeight, 100U,
+                     osr::conditional_numeric_unit::kWeight100Kg},
+           test_case{"maxweight:hgv:conditional"sv, "10 cwt"sv,
+                     osr::conditional_restriction_field::kMaxWeight, 5U,
                      osr::conditional_numeric_unit::kWeight100Kg},
            test_case{"maxaxleload:conditional"sv, "20000 lb"sv,
                      osr::conditional_restriction_field::kMaxAxleLoad, 91U,
